Unify provider configuration naming

+++ see/image_providers/os_glance.py
@@ -18,14 +18,14 @@ it doesn't already exist on the configured target path. Images can be requested
 by name or UUID; if name is requested the latest matching image is retrieved.
 
 provider_parameters:
-    target_path (str): Absolute path where to download the image. If target_path
-                       is an existing file, it will be overwritten if the image
-                       is newer. Otherwise target_path is understood to be a
-                       directory and the image's UUID will be used as filename.
-    os_auth (dict):    A dictionary with OpenStack authentication parameters as
-                       needed by OpenStack's Keystone client.
-    session (dict):    A dictionary with OpenStack Session parameters. Allows
-                       authentication to Keystone over TLS.
+    path (str):     Absolute path where to download the image. If target_path
+                    is an existing file, it will be overwritten if the image
+                    is newer. Otherwise target_path is understood to be a
+                    directory and the image's UUID will be used as filename.
+    os_auth (dict): A dictionary with OpenStack authentication parameters as
+                    needed by OpenStack's Keystone client.
+    session (dict): A dictionary with OpenStack Session parameters. Allows
+                    authentication to Keystone over TLS.
 
 """
 
@@ -65,30 +65,28 @@ class GlanceProvider(ImageProvider):
         try:
             metadata = self._retrieve_metadata()
         except FileNotFoundError:
-            if os.path.exists(self.configuration['target_path']):
-                if os.path.isfile(os.path.realpath(
-                        self.configuration['target_path'])):
-                    return self.configuration['target_path']
+            if os.path.exists(self.configuration['path']):
+                if os.path.isfile(os.path.realpath(self.configuration['path'])):
+                    return self.configuration['path']
                 else:
                     for image in self._find_potentials():
-                        tgt = os.path.join(self.configuration['target_path'],
-                                           image.id)
+                        tgt = os.path.join(self.configuration['path'], image.id)
                         if os.path.exists(tgt):
                             return tgt
             raise
 
-        if (os.path.exists(self.configuration['target_path']) and
+        if (os.path.exists(self.configuration['path']) and
                 os.path.isfile(os.path.realpath(
-                    self.configuration['target_path'])) and
+                    self.configuration['path'])) and
                 datetime.fromtimestamp(os.path.getmtime(
-                    self.configuration['target_path'])) >
+                    self.configuration['path'])) >
                 datetime.strptime(metadata.updated_at, "%Y-%m-%dT%H:%M:%SZ")):
-            return self.configuration['target_path']
+            return self.configuration['path']
 
-        target = (self.configuration['target_path']
-                  if os.path.isfile(self.configuration['target_path'])
-                  else '/'.join((self.configuration['target_path'].rstrip('/'),
-                                 metadata.id)))
+        target = (self.configuration['path']
+                  if os.path.isfile(self.configuration['path'])
+                  else os.path.join(self.configuration['path'], metadata.id))
+
         os.makedirs(os.path.dirname(os.path.realpath(target)))
 
         self._download_image(metadata, target)
@@ -127,26 +125,28 @@ class GlanceProvider(ImageProvider):
 
     def _find_potentials(self):
         return sorted([image for image in self.glance_client.images.list()
-                       if (image.id == self.uri or image.name == self.uri)
+                       if (image.id == self.name or image.name == self.name)
                        and image.status != 'active'],
                       key=lambda x: x.updated_at, reverse=True)
 
     def _retrieve_metadata(self):
         try:
             return sorted([image for image in self.glance_client.images.list()
-                           if (image.id == self.uri or image.name == self.uri)
+                           if (image.id == self.name or image.name == self.name)
                            and image.status == 'active'],
                           key=lambda x: x.updated_at, reverse=True)[0]
         except IndexError:
-            raise FileNotFoundError(self.uri)
+            raise FileNotFoundError(self.name)
 
     def _download_image(self, img_metadata, target):
-        img_downloader = self.glance_client.images.data(img_metadata.id)
-        _, temp = tempfile.mkstemp(dir=os.path.dirname(target), suffix='.part')
-        with open(temp, 'wb') as imagefile:
-            for chunk in img_downloader:
-                imagefile.write(chunk)
-        if not verify_checksum(temp, img_metadata.checksum):
-            os.remove(temp)
-            raise RuntimeError('Checksum failure. File: %s' % target)
-        os.rename(temp, target)
+        if not os.path.exists(target):
+            img_downloader = self.glance_client.images.data(img_metadata.id)
+            _, temp = tempfile.mkstemp(dir=os.path.dirname(target),
+                                       suffix='.part')
+            with open(temp, 'wb') as imagefile:
+                for chunk in img_downloader:
+                    imagefile.write(chunk)
+            if not verify_checksum(temp, img_metadata.checksum):
+                os.remove(temp)
+                raise RuntimeError('Checksum failure. File: %s' % target)
+            os.rename(temp, target)
